+++ Landoria.HammerFreedom/HammerFreedomAuthorization.cs
@@ -129,7 +129,8 @@ namespace Landoria.HammerFreedom
                 hammerWorld, settings != null && settings.Flight,
                 settings != null && settings.FallDamageImmunity,
                 settings != null && settings.UnlimitedStamina,
-                settings != null && settings.NoDurabilityLoss);
+                settings != null && settings.NoDurabilityLoss,
+                settings != null && settings.RecoverBuildMaterials);
         }
 
         private static void ResetConnection()